feat(consensus): mock_chain_validation policy for non-validating replay builds
Adds a build-tag-gated ConsensusPolicy ("mock_chain_validation") so a
non-validating node can replay real chain history despite the consensus
divergence inherent to replaying an AppHash-breaking storage migration
(memIAVL->flatKV) against a validator set it cannot reproduce bit-for-bit.

The default build is unchanged: DefaultConsensusPolicy().HandleError preserves
every panic/halt, verified byte-identical.

Under the mock_chain_validation tag the policy swallows an explicit ALLOWLIST of
the sentinels that drift for migration / validator-set / commit reasons. Peer-
supplied block-content integrity still HALTS -- ErrDataHash (the tx merkle root
feeding the EVM execution this build compares), ErrEvidenceHash,
ErrPerEvidenceValidateBasic -- so a malformed/lying peer cannot poison the audit
input. Allowlist => a sentinel added later halts by default.

Previously-unguarded checks route through the policy (inert in the default build),
all via the single HandleError(err) verb:
- blocksync/reactor.go: VerifyCommitLight -> policy.HandleError
- state/execution.go: buildLastCommitInfo routes a commit/validator-set size
  mismatch through the policy (default panics; mock_chain_validation builds
  best-effort commit info, which only feeds staking rewards/downtime)
- statesync/syncer.go: state-sync snapshot appHash check -> policy (also extends
  mock_block_validation, which already relaxes ErrAppHash)
- x/upgrade/abci.go: binary-updated-before-trigger panic -> policy

Tests: SwallowMatrix pins the allowlist + the never-swallow content-integrity
guard; TestBlockValidateBasic probes the active policy; a build-tagged test pins
buildLastCommitInfo's best-effort path (commit/valset size mismatch, no panic).
