+++ sei-cosmos/x/upgrade/abci_test.go
@@ -121,40 +121,6 @@ func VerifyDoUpgradeWithCtx(t *testing.T, newCtx sdk.Context, proposalName strin
 	VerifyCleared(t, newCtx)
 }
 
-func TestHaltIfTooNew(t *testing.T) {
-	s := setupTest(t, 10)
-	t.Log("Verify that we don't panic with registered plan not in database at all")
-	var called int
-	s.keeper.SetUpgradeHandler("future", func(_ sdk.Context, _ types.Plan, vm module.VersionMap) (module.VersionMap, error) {
-		called++
-		return vm, nil
-	})
-
-	newCtx := s.ctx.WithBlockHeight(s.ctx.BlockHeight() + 1).WithBlockTime(time.Now())
-	require.NotPanics(t, func() {
-		upgrade.BeginBlocker(s.keeper, newCtx)
-	})
-	require.Equal(t, 0, called)
-
-	t.Log("Verify we panic if we have a registered handler ahead of time")
-	err := s.handler(s.ctx, &types.SoftwareUpgradeProposal{Title: "prop", Plan: types.Plan{Name: "future", Height: s.ctx.BlockHeight() + 3}})
-	require.NoError(t, err)
-	require.Panics(t, func() {
-		upgrade.BeginBlocker(s.keeper, newCtx)
-	})
-	require.Equal(t, 0, called)
-
-	t.Log("Verify we no longer panic if the plan is on time")
-
-	futCtx := s.ctx.WithBlockHeight(s.ctx.BlockHeight() + 3).WithBlockTime(time.Now())
-	require.NotPanics(t, func() {
-		upgrade.BeginBlocker(s.keeper, futCtx)
-	})
-	require.Equal(t, 1, called)
-
-	VerifyCleared(t, futCtx)
-}
-
 func VerifyCleared(t *testing.T, newCtx sdk.Context) {
 	t.Log("Verify that the upgrade plan has been cleared")
 	bz, err := s.querier(newCtx, []string{types.QueryCurrent}, abci.RequestQuery{})
