+++ sei-tendermint/internal/consensus/state_test.go
@@ -201,53 +201,6 @@ func TestStateEnterProposeYesPrivValidator(t *testing.T) {
 	ensureNoNewTimeout(t, timeoutCh, cs.state.ConsensusParams.Timeout.ProposeTimeout(round).Nanoseconds())
 }
 
-func TestStateBadProposal(t *testing.T) {
-	config := configSetup(t)
-	ctx := t.Context()
-
-	cs1, vss := makeState(ctx, t, makeStateArgs{config: config, validators: 2})
-	height, round := cs1.roundState.Height(), cs1.roundState.Round()
-	vs2 := vss[1]
-
-	partSize := types.BlockPartSizeBytes
-	proposalCh := subscribe(ctx, t, cs1.eventBus, types.EventQueryCompleteProposal)
-	voteCh := subscribe(ctx, t, cs1.eventBus, types.EventQueryVote)
-
-	propBlock, err := cs1.createProposalBlock(ctx)
-	require.NoError(t, err)
-
-	round++
-	incrementRound(vss[1:]...)
-
-	stateHash := propBlock.AppHash
-	if len(stateHash) == 0 {
-		stateHash = make([]byte, 32)
-	}
-	stateHash[0] = (stateHash[0] + 1) % 255
-	propBlock.AppHash = stateHash
-	propBlockParts, err := propBlock.MakePartSet(partSize)
-	require.NoError(t, err)
-	blockID := types.BlockID{Hash: propBlock.Hash(), PartSetHeader: propBlockParts.Header()}
-	pubKey, err := vss[1].PrivValidator.GetPubKey(ctx)
-	require.NoError(t, err)
-	proposal := types.NewProposal(vs2.Height, round, -1, blockID, propBlock.Header.Time, propBlock.GetTxHashes(), propBlock.Header, propBlock.LastCommit, propBlock.Evidence, pubKey.Address())
-	p := proposal.ToProto()
-	require.NoError(t, vs2.SignProposal(ctx, config.ChainID(), p))
-	proposal.Signature = utils.OrPanic1(crypto.SigFromBytes(p.Signature))
-
-	err = cs1.SetProposalAndBlock(ctx, proposal, propBlock, propBlockParts, "some peer")
-	require.NoError(t, err)
-
-	cs1.startTestRound(ctx, height, round)
-	ensureProposal(t, proposalCh, height, round, blockID)
-	ensurePrevoteMatch(t, voteCh, height, round, nil)
-	cs1.signAddVotes(ctx, t, tmproto.PrevoteType, config.ChainID(), blockID, vs2)
-	ensurePrevote(t, voteCh, height, round)
-	ensurePrecommit(t, voteCh, height, round)
-	cs1.validatePrecommit(ctx, t, round, -1, vss[0], nil, nil)
-	cs1.signAddVotes(ctx, t, tmproto.PrecommitType, config.ChainID(), blockID, vs2)
-}
-
 func TestStateOversizedBlock(t *testing.T) {
 	config := configSetup(t)
 	ctx := t.Context()
