+++ sei-tendermint/types/consensus_policy_mock_block_validation_test.go
@@ -24,6 +24,7 @@ func TestConsensusPolicy_MockBlockValidation_Matrix(t *testing.T) {
 		ErrLastCommitHash:            false,
 		ErrEvidenceHash:              false,
 		ErrPerEvidenceValidateBasic:  false,
+		ErrUpgradeBeforeTrigger:      true,
 	}
 	for _, sentinel := range ValidationErrors() {
 		swallow, ok := swallowExpected[sentinel]
