+++ sei-tendermint/types/consensus_policy_mock_chain_validation_test.go
@@ -10,18 +10,30 @@ import (
 
 func TestConsensusPolicy_MockChainValidation_SwallowMatrix(t *testing.T) {
 	policy := DefaultConsensusPolicy()
-	for _, sentinel := range ValidationErrors() {
-		// A contextual error wrapping the sentinel must match it under errors.Is.
-		err := fmt.Errorf("validation failed: %w", sentinel)
-		got := policy.HandleError(err)
-		if sentinel == ErrLastCommitVerify {
-			if got != err {
-				t.Errorf("mock_chain_validation ConsensusPolicy.HandleError(%v) = %v, want the input error (excluded from swallow set)", sentinel, got)
-			}
-			continue
+
+	swallowed := make(map[error]bool, len(swallowedErrors))
+	for _, e := range swallowedErrors {
+		swallowed[e] = true
+	}
+
+	// Semantic guard: peer-supplied block-content integrity must NEVER be in the
+	// swallow allowlist, so a malformed/lying peer cannot silently poison the
+	// audit input. This is the load-bearing safety property of this build.
+	for _, e := range []error{ErrDataHash, ErrEvidenceHash, ErrPerEvidenceValidateBasic} {
+		if swallowed[e] {
+			t.Errorf("%v is in the swallow allowlist; peer-supplied content integrity must halt", e)
 		}
-		if got != nil {
-			t.Errorf("mock_chain_validation ConsensusPolicy.HandleError(%v) = %v, want nil", sentinel, got)
+	}
+
+	// HandleError swallows exactly the allowlist; everything else in the audit set
+	// halts, including any newly added sentinel (the allowlist is halt-by-default).
+	for _, sentinel := range ValidationErrors() {
+		got := policy.HandleError(fmt.Errorf("validation failed: %w", sentinel))
+		switch {
+		case swallowed[sentinel] && got != nil:
+			t.Errorf("HandleError(%v) = %v, want swallowed (nil)", sentinel, got)
+		case !swallowed[sentinel] && got == nil:
+			t.Errorf("HandleError(%v) = nil, want HALT (not in the swallow allowlist)", sentinel)
 		}
 	}
 }
